pyproject.toml
[project]
name = "graph-based-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=2.0.0a3",
]
